Split up the module. Use simple types over classes
