+++ lib/types.ts
@@ -1,0 +1,1 @@
+export type BooleanNumber = 1 | 0;
